Fix 'new project' command for Ruby 3.4.x versions
As discussed in issue #363, the `terraspace new project` command was not working with ruby >= 3.4.X.
The main issue being the way line tracing is done in the call stack in the
Terraspace::Command#subcommand? method.

For Ruby 3.3.x the code is looking for "in subcommand" substring and matching on this line:

  e.g., "thor.rb:338:in `block in subcommand'"

However for Ruby 3.4.x this line is now:

  e.g., "thor.rb:338:in 'block in Terraspace::CLI#subcommand'"

The fix here is to use l.match? and a regex instead of l.include? to match both of these lines from
either Ruby version.